feat: 安全扫描工具 - 漏洞/敏感信息/权限扫描
- SecurityScanTool: 4种扫描模式
- vulnerabilities: 代码安全漏洞检测
- secrets: 敏感信息扫描（API Key/Token等）
- permissions: 文件权限检查
- full: 目录完整扫描
- 7个单元测试
- 433个测试全部通过

+++ src/DesktopMascot.Core/Caching/MemoryCache.cs
@@ -136,8 +136,8 @@ public class MemoryCache : ICache
         return new CacheStatistics
         {
             TotalEntries = entries.Count,
-            TotalHits = Interlocked.Read(ref _hits),
-            TotalMisses = Interlocked.Read(ref _misses),
+            TotalHits = (int)Interlocked.Read(ref _hits),
+            TotalMisses = (int)Interlocked.Read(ref _misses),
             TotalSizeBytes = entries.Sum(e => e.SizeBytes),
             OldestEntry = entries.Any() ? entries.Min(e => e.CreatedAt) : null,
             NewestEntry = entries.Any() ? entries.Max(e => e.CreatedAt) : null
